refactor: limit public api, refactor mapping methods, restructure grpc errors dir
