+++ src/errors/index.ts
@@ -1,8 +1,10 @@
 /**
  * Public entrypoint for the transport-agnostic A2A error hierarchy.
- * Explicit named re-exports so this file defines the API contract of
- * `@a2a-js/sdk/errors`; a `base.ts` helper added tomorrow does NOT
- * accidentally become public API.
+ * Contract of `@a2a-js/sdk/errors`: semantic error classes, transport
+ * variants, type guards, wire helpers, and the constants a caller
+ * needs to construct or interpret them. Internal registries (spec
+ * tables, class maps) stay in `./base.ts` / `./rest.ts` / `./json_rpc.ts`
+ * and are consumed only inside `src/`.
  *
  * gRPC errors live at `@a2a-js/sdk/errors/grpc` because their
  * encode/decode helpers pull `@bufbuild/protobuf`.
@@ -11,14 +13,9 @@
 // --- base ---
 export {
   A2A_ERROR_DOMAIN,
-  A2A_ERROR_CLASSES,
-  A2A_ERROR_SPECS,
-  A2A_ERROR_SPECS_BY_REASON,
   A2AError,
-  type A2AErrorClass,
   type A2AErrorInfo,
   type A2AErrorOptions,
-  type A2AErrorSpec,
   ContentTypeNotSupportedError,
   ERROR_INFO_TYPE,
   type ErrorDetail,
@@ -28,7 +25,6 @@ export {
   InvalidAgentResponseError,
   PushNotificationNotSupportedError,
   RequestMalformedError,
-  specForError,
   TaskNotCancelableError,
   TaskNotFoundError,
   UnsupportedOperationError,
@@ -37,27 +33,23 @@ export {
 
 // --- REST transport ---
 export {
+  fromRestErrorBody,
   HTTP_STATUS,
   isRestError,
-  REST_ERROR_CLASSES,
-  REST_ERROR_HTTP_STATUS,
-  REST_ERROR_STATUS_NAME,
-  REST_STATUS_NAME,
   type RestA2AError,
   type RestA2AErrorOptions,
-  type RestErrorBody,
   RestContentTypeNotSupportedError,
+  type RestErrorBody,
   RestExtendedAgentCardNotConfiguredError,
   RestExtensionSupportRequiredError,
   RestInvalidAgentResponseError,
   RestPushNotificationNotSupportedError,
   RestRequestMalformedError,
+  restStatusFor,
   RestTaskNotCancelableError,
   RestTaskNotFoundError,
   RestUnsupportedOperationError,
   RestVersionNotSupportedError,
-  fromRestErrorBody,
-  restStatusFor,
   toRestErrorBody,
 } from './rest.js';
 
@@ -66,9 +58,6 @@ export {
   A2A_ERROR_CODE,
   fromJsonRpcErrorResponse,
   isJsonRpcError,
-  JSON_RPC_CODE_TO_ERROR,
-  JSON_RPC_ERROR_CLASSES,
-  JSON_RPC_ERROR_CODE,
   type JsonRpcA2AError,
   type JsonRpcA2AErrorOptions,
   JsonRpcContentTypeNotSupportedError,
